Allow for CIDR notation in IP exclude list
This would allow user to exclude their entire local subnet (e.g. - "192.168.1.0/24"). Also, I have a lot of entries from 108.177.70.0/24 and 108.177.71.0/24, which belong to Google and I assume are due to Assistant integration (not using Nabu Casa). CIDR notation is a lot cleaner than adding each of the 510 addresses in those ranges...